Move Format prefixed exceptions into their own file
After writig encoding/1 in the Format module it was obviously annoying
to have the exceptions listed before the module's main functions.
